pyproject.toml
[project]
name = "multiplayer-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.3.0",
    "jupyterlab>=4.4.0",
    "logfire>=3.14.0",
    "mermaid-python>=0.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pydantic>=2.11.1",
    "pydantic-ai>=0.1.3",
    "pydantic-graph>=0.1.3",
    "pygame==2.6.1",
    "pygame-widgets>=1.1.7",
    "python-dotenv>=1.0.1",
    "rich>=14.0.0",
]
